requirements.txt
pyspark==3.5.3
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
kafka-python==2.0.2
